reborn/requirements.txt
altgraph==0.17
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
click==8.0.1
colorama==0.4.4
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
filelock==3.6.0
Flask==1.1.2
future==0.18.2
idna==2.10
importlib-metadata==4.11.3
itsdangerous==2.0.1
Jinja2==2.10.3
Markdown==3.3.6
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
mysqlclient==2.0.3
npm==0.1.1
npx==0.1.1
numpy==1.22.3
optional-django==0.1.0
pefile==2019.4.18
Pillow==9.1.0
platformdirs==2.5.1
protobuf==3.20.0
pycparser==2.21
pyinstaller @ https://github.com/pyinstaller/pyinstaller/archive/develop.zip
pyinstaller-hooks-contrib==2021.1
python3-memcached==1.51
pytz==2021.1
pywin32-ctypes==0.2.0
pyzmq==22.0.3
requests==2.25.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.1
tzdata==2021.5
urllib3==1.26.5
virtualenv==20.14.0
Werkzeug==1.0.1
zipp==3.8.0
zmq==0.0.0
